requirements.txt
asgiref==3.4.1
Django==3.2.5
django-environ==0.4.5
django-filter==2.4.0
django-shortuuidfield==0.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Faker==8.10.1
pendulum==2.1.2
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
pytzdata==2020.1
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
